+++ pkg/autometrics/instrument.go
@@ -1,14 +1,10 @@
 package autometrics
 
 import (
+	"reflect"
 	"runtime"
 	"strings"
 )
-
-type Option interface {
-	// Apply the option to the currently created context
-	Apply(*Context)
-}
 
 // CallerInfo returns the (method name, module name) of the function that called the function that called this function.
 //
@@ -37,7 +33,7 @@ func CallerInfo() (callInfo CallInfo) {
 	index := strings.LastIndex(functionName, ".")
 
 	if index == -1 {
-		callInfo.FuncName = frame.Func.Name()
+		callInfo.FuncName = functionName
 	} else {
 		moduleIndex := strings.LastIndex(functionName[:index], ".")
 		if moduleIndex == -1 {
@@ -60,7 +56,7 @@ func CallerInfo() (callInfo CallInfo) {
 	index = strings.LastIndex(parentFunctionName, ".")
 
 	if index == -1 {
-		callInfo.ParentFuncName = parentFrame.Func.Name()
+		callInfo.ParentFuncName = parentFunctionName
 	} else {
 		moduleIndex := strings.LastIndex(parentFunctionName[:index], ".")
 		if moduleIndex == -1 {
@@ -73,4 +69,28 @@ func CallerInfo() (callInfo CallInfo) {
 	}
 
 	return
+}
+
+// ReflectFunctionModuleName takes any function and returns it's name and module split.
+//
+// There is no `caller` in this context (we just use reflection to extract the information
+// from the function pointer), therefore the caller-related fields in the return value are
+// empty.
+func ReflectFunctionModuleName(f interface{}) (callInfo CallInfo) {
+	functionName := runtime.FuncForPC(reflect.ValueOf(f).Pointer()).Name()
+
+	index := strings.LastIndex(functionName, ".")
+	if index == -1 {
+		callInfo.FuncName = functionName
+	} else {
+		moduleIndex := strings.LastIndex(functionName[:index], ".")
+		if moduleIndex == -1 {
+			callInfo.ModuleName = functionName[:index]
+		} else {
+			callInfo.ModuleName = functionName[moduleIndex+1 : index]
+		}
+		callInfo.FuncName = functionName[index+1:]
+	}
+
+	return callInfo
 }